change return type from BigDecimal to long of overlappingDays()

+++ Budget/src/main/java/budget/BudgetService.java
@@ -30,9 +30,9 @@ public class BudgetService {
             return yearMonth.compareTo(startYearMonth) >= 0 && yearMonth.compareTo(endYearMonth) <= 0;
         }).map(budget -> {
 
-            BigDecimal days = new Period(start, end).overlappingDays(budget);
+            long days = new Period(start, end).overlappingDays(budget);
             BigDecimal dailyAmount = budget.getAmount().divide(new BigDecimal(budget.getYearMonthInstance().lengthOfMonth()), 0, RoundingMode.HALF_UP);
-            return dailyAmount.multiply(days);
+            return dailyAmount.multiply(BigDecimal.valueOf(days));
         }).reduce(BigDecimal.ZERO, BigDecimal::add);
     }
 }
